requirements.txt
apache-airflow==2.10.5
apache-airflow-providers-mongo==5.0.2
requests 
